rename Madam.{c,h} to madam.{c,h}
