+++ libfreedo/madam.c
@@ -27,7 +27,7 @@ Felix Lazarev
 
 #include <retro_inline.h>
 
-#include "Madam.h"
+#include "madam.h"
 #include "clio.h"
 #include "vdlp.h"
 #include "arm.h"
