+++ libfreedo/madam.h
